Update Getting-Started.md
First batch of changes for Getting Started. Since this guide is quite long and complex, I am going to edit it in two batches. After we close on this first batch of edits and merge them, I'll pick up from ## Generating a client-side Auditing Library and create the second batch of edits.